[refactor] Move all "untravel" functions to new module
