+++ packages/prosemirror-lwdita/src/document.ts
@@ -269,286 +269,6 @@ export function document(jdita: JDita): Record<string, any> {
   throw new Error('jdita must be a document');
 }
 
-/**
- * Replace underscores with hyphens in node names
- *
- * @param type - The string to be modified
- * @returns The sanitized node name with hyphens
- */
-function getJditaNodeName(type: string): string {
-  return type.replace(/_/g, '-');
-}
-
-/**
-* Recursively traverse through all items in the Prosemirror DOM
-* and create a JDITA object
-*
-* @param prosemirrorDocument - The Prosemirror DOM object
-* @returns The JDITA object
-*/
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function unTravel(prosemirrorDocument: Record<string, any>): JDita {
-
-  // Prosemirror content will become JDITA children
-  const children = prosemirrorDocument.content?.map(unTravel);
-
-  // attrs will become attributes
-  const attributes = prosemirrorDocument.attrs || {};
-
-  // get the node name
-  const nodeName = getJditaNodeName(prosemirrorDocument.type);
-
-  if (nodeName === 'video' || nodeName === 'audio' || nodeName === 'image') {
-    return mediaNodeUntravel(nodeName, attributes, children)
-  }
-
-  // handle the attributes
-  delete attributes['parent'];
-  for (const key in attributes) {
-    if (!attributes[key]) {
-      delete attributes[key];
-    }
-  }
-
-  if(nodeName === 'hard-break') {
-    return {
-      nodeName: 'text',
-      content: '\n'
-    }
-  }
-
-  if (nodeName === 'text') {
-    // check for marks and wrap the content in the mark node
-
-    if (prosemirrorDocument.marks) {
-      const mark = prosemirrorDocument.marks[0].type;
-      return {
-        nodeName: mark,
-        attributes: attributes,
-        children: [
-          {
-            nodeName: "text",
-            content: prosemirrorDocument.text
-          }
-        ]
-      }
-    }
-    return {
-      nodeName,
-      'content': prosemirrorDocument.text
-    }
-  }
-
-  const nodeObject: JDita = {
-    nodeName,
-    attributes,
-    children
-  }
-
-  return nodeObject;
-}
-
-/**
- * Special untravel function for media nodes.
- * Reverses the transformation done by the NODES[value.nodeName] in the travel function
- *
- * @param nodeName - string node name
- * @param attributes - node attributes
- * @param children - JDita[] node children
- * @param prosemirrorDocument - prosemirror document
- * @returns JDita node
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function mediaNodeUntravel(nodeName: string, attributes: Record<string, string>, children: JDita[]): JDita {
-  if (nodeName === 'video') {
-    // we must populate the video node with the necessary attributes and children
-    const allAttributes = {
-      props: attributes.props,
-      dir: attributes.dir,
-      "xml:lang": attributes['xml:lang'],
-      translate: attributes.translate,
-      href: attributes.href,
-      format: attributes.format,
-      scope: attributes.scope,
-      id: attributes.id,
-      conref: attributes.conref,
-      outputclass: attributes.outputclass,
-      class: attributes.class,
-      width: attributes.width,
-      height: attributes.height,
-      autoplay: attributes.autoplay,
-      controls: attributes.controls,
-      loop: attributes.loop,
-      muted: attributes.muted,
-      tabindex: attributes.tabindex,
-    }
-
-    const allChildren: JDita[] = [];
-    //children[0] resembles the video desc this value does not change
-    allChildren.push(children[0]) // video desc node
-
-    if (attributes.title !== undefined) {
-      const desc: JDita = createMediaChild('desc', attributes.title);
-      allChildren.push(desc);
-    }
-
-    if (attributes.fallback !== undefined) {
-      const fallback: JDita = createMediaChild('fallback', attributes.fallback);
-      allChildren.push(fallback);
-    }
-
-    if (attributes.poster !== undefined) {
-      const poster: JDita = createMediaChild('video-poster', attributes.poster);
-      allChildren.push(poster);
-    }
-
-    if (attributes.track !== undefined) {
-      const track: JDita = createMediaChild('media-track', attributes.track);
-      allChildren.push(track);
-    }
-
-    if (attributes.source !== undefined) {
-      const source: JDita = createMediaChild('media-source', attributes.source);
-      allChildren.push(source);
-    }
-
-    allChildren.push(children[1])
-
-    // return the created video node
-    return {
-      nodeName,
-      'attributes': allAttributes,
-      'children': allChildren
-    }
-  }
-
-  if (nodeName === 'audio') {
-    const allAudioAttributes = {
-      class: attributes.class,
-      conref: attributes.conref,
-      "xml:lang": attributes['xml:lang'],
-      dir: attributes.dir,
-      id: attributes.id,
-      outputclass: attributes.outputclass,
-      props: attributes.props,
-      translate: attributes.translate,
-      autoplay: attributes.autoplay,
-      controls: attributes.controls,
-      loop: attributes.loop,
-      muted: attributes.muted,
-      source: attributes.source,
-      href: attributes.href,
-      format: attributes.format,
-      scope: attributes.scope,
-      tabindex: attributes.tabindex,
-    }
-
-    const allAudioChildren: JDita[] = [];
-    allAudioChildren.push(children[0])
-
-    if (attributes.desc !== undefined) {
-      const desc: JDita = createMediaChild('desc', attributes.desc);
-      allAudioChildren.push(desc);
-    }
-
-    if (attributes.fallback !== undefined) {
-      const fallback: JDita = createMediaChild('fallback', attributes.fallback);
-      allAudioChildren.push(fallback);
-    }
-
-    if (attributes.source !== undefined) {
-      const source: JDita = createMediaChild('media-source', attributes.source);
-      allAudioChildren.push(source);
-    }
-
-    if (attributes.track !== undefined) {
-      const track: JDita = createMediaChild('media-track', attributes.track);
-      allAudioChildren.push(track);
-    }
-
-    allAudioChildren.push(children[1])
-
-    return {
-      nodeName,
-      'attributes': allAudioAttributes,
-      'children': allAudioChildren
-    }
-  }
-
-  if (nodeName === 'image') {
-    const allImageAttributes = {
-      // %localization
-      dir: attributes.dir,
-      "xml:lang": attributes['xml:lang'],
-      translate: attributes.translate,
-      // %reference-content
-      href: attributes.href,
-      format: attributes.format,
-      scope: attributes.scope,
-      // class attributes
-      outputclass: attributes.outputclass,
-      class: attributes.class,
-      // custom attributes
-      keyref: attributes.keyref,
-      width: attributes.width,
-      height: attributes.height,
-    }
-
-    const allImageChildren: JDita[] = [];
-    allImageChildren.push({
-      nodeName: 'alt',
-      // all of the attributes will be undefined as the alt node only contains text
-      attributes: {
-        dir: undefined,
-        "xml:lang": undefined,
-        translate: undefined,
-        keyref: undefined,
-        outputclass: undefined,
-        class: undefined
-      },
-      // FIXME: allowed children of `alt` are: 'b', 'em', 'i', 'ph', 'strong', 'sub', 'sup', 'tt', 'u'
-      children: [
-        {
-          nodeName: 'text',
-          content: attributes.alt
-        }
-      ]
-    })
-
-    return {
-      nodeName,
-      'attributes': allImageAttributes,
-      'children': allImageChildren
-    }
-  }
-  throw new Error('Invalid node name');
-}
-
-/**
- * Creates children for media nodes
- * Children like media-autoplay, media-controls, media-loop, media-muted, video-poster, media-source share all the same structure
- * This is a helper function to create these children
- *
- * @param nodeName - string
- * @param value - string
- * @returns media child node
- */
-function createMediaChild(nodeName: string,value: string): JDita {
-  return {
-    nodeName: nodeName,
-    attributes: {
-      dir: undefined,
-      "xml:lang": undefined,
-      translate: undefined,
-      name: undefined,
-      href: value,
-      outputclass: undefined,
-      class: undefined,
-    },
-    children: undefined
-  };
-}
-
 //Escape hatch for Unit Testing due to a lack of “package-private” accessibility scope in TypeScript
 export const _test_private_document = {
   deleteUndefined,
