+++ packages/prosemirror-lwdita/src/tests/document.spec.ts
@@ -17,7 +17,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
 import ChaiPromised from 'chai-as-promised';
 import { use, expect, assert } from 'chai';
-import { document, NODES, _test_private_document, unTravel } from '../document';
+import { document, NODES, _test_private_document } from '../document';
+import { unTravel } from '../untravelNodes';
 import {
   JDITA_OBJECT,
   TRANSFORMED_JDITA_OBJECT,
